chore(tests): move factory traits to factory namespace
