+++ tests/Concern/Factory/BoxFactoryTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Concern;
+namespace App\Tests\Concern\Factory;
 
 use App\Siklid\Document\Box;
 use App\Siklid\Document\User;
