+++ tests/Concern/Factory/FactoryConcern.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Concern;
+namespace App\Tests\Concern\Factory;
 
+use App\Tests\Concern\WithFaker;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 trait FactoryConcern
